Add files via upload
Scenario
A security incident affecting a company's website yummyrecipesforme.com. A former employee used brute-force attack to gain access to the administrative dashboard. The paasword to the administrative account was not changed at was using the default password which made it easy for the attacker to gain access. The attacker edited the website's source-code and added malicious javascript file that made visitors download and executable. The browser redirected them to a fake website containing malware which they later reported that it made their devices slow.
My task was to investigate the incident using provided tcpdump logs, identify the network protocol(s) involved in the traffic, and document the event in an incident report. I was also required to provide one recommendation that could help prevent brute-force attacks in the future.
My Response
  I reviewed the packet capture and identified DNS and HTTP traffic. DNS was used to resolve domain names to IP addresses, and HTTP handled webpage requests and the malware download.
  I wrote a report summarizing how the attacker accessed the admin panel, how the malicious script functioned, and how users were redirected to the fake website. I included evidence from the tcpdump logs and sandbox analysis.